tablestore-java-mcp-server-rag/knowledge-data-generator/requirements.txt
mcp==1.5.0
dashscope==1.22.2
openai
markdown-it-py
markdownify
pdfplumber
comtypes